requirements.txt
SpeechRecognition==3.10.0
django-environ==0.11.2
gtts==2.4.0
pyTelegramBotAPI==4.14.0
pydub==0.25.1